separate streaming to own file
